vagrant/catalog/requirements.txt
Flask-SqlAlchemy
Flask-OAuthlib
Flask-Session
